feat: create post

+++ internal/models/requests.go
@@ -2,7 +2,7 @@ package models
 
 // Transacao representa a estrutura da tabela historico_transacoes
 type PostTransacaoRequest struct {
-	Valor     float64 `json:"valor"`
-	Tipo      string  `json:"tipo"`
-	Descricao string  `json:"descricao"`
+	Valor     int64  `json:"valor"`
+	Tipo      string `json:"tipo"`
+	Descricao string `json:"descricao"`
 }
